Replace double with boost cpp_int for dist. calc.

+++ C++/kd-map/test_kdmapdynamic.cpp
@@ -382,7 +382,7 @@ int main(int argc, char** argv) {
 
       // Find numNeighbors nearest neighbors to each coordinate.
       if (neighbors) {
-        forward_list< pair<double, KdNode<kdKey_t, kdValue_t>*> > neighborList;
+        forward_list< pair<cpp_int, KdNode<kdKey_t, kdValue_t>*> > neighborList;
         vector<kdKey_t> query(numDimensions);
         auto beginTime = steady_clock::now();
         for (size_t i = 0; i < coordinates.size(); ++i) {
@@ -450,7 +450,7 @@ int main(int argc, char** argv) {
 
     // Find numNeighbors nearest neighbors to each coordinate.
     if (neighbors) {
-      forward_list< pair<double, KdNode<kdKey_t, kdValue_t>*> > neighborList;
+      forward_list< pair<cpp_int, KdNode<kdKey_t, kdValue_t>*> > neighborList;
       vector<kdKey_t> query(numDimensions);
       auto beginTime = steady_clock::now();
       for (size_t i = 0; i < coordinates.size(); ++i) {
